Relocated ContentManagerModel::Downloads
A new class DownloadManager was introduced as a new home for
ContentManagerModel::Downloads, but it is going to grow into a more
serious unit of code that will consolidate most functionality
related to download management.
