+++ src/contentmanagermodel.h
@@ -5,10 +5,9 @@
 #include <QModelIndex>
 #include <QVariant>
 #include <QIcon>
-#include <QMutex>
-#include <QMutexLocker>
 #include "thumbnaildownloader.h"
 #include "rownode.h"
+#include "downloadmanagement.h"
 #include <memory>
 
 class ContentManager;
@@ -24,39 +23,7 @@ public: // types
     typedef QMap<QString, QVariant> BookInfo;
     typedef QList<BookInfo>         BookInfoList;
 
-    // BookId -> DownloadState map
-    class Downloads
-    {
-    private:
-        typedef std::shared_ptr<DownloadState> DownloadStatePtr;
-        typedef QMap<QString, DownloadStatePtr> ImplType;
-
-    public:
-        void set(const QString& id, DownloadStatePtr d) {
-            const QMutexLocker threadSafetyGuarantee(&mutex);
-            impl[id] = d;
-        }
-
-        DownloadStatePtr value(const QString& id) const {
-            const QMutexLocker threadSafetyGuarantee(&mutex);
-            return impl.value(id);
-        }
-
-        QList<QString> keys() const {
-            const QMutexLocker threadSafetyGuarantee(&mutex);
-            return impl.keys();
-        }
-
-        void remove(const QString& id) {
-            const QMutexLocker threadSafetyGuarantee(&mutex);
-            impl.remove(id);
-        }
-
-    private:
-        ImplType impl;
-        mutable QMutex mutex;
-    };
-
+    typedef DownloadManager::Downloads Downloads;
 
 public: // functions
     explicit ContentManagerModel(ContentManager* contentMgr);
